+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Jitrapon Tiachunpun
 2. Fix question text colored as black in dark mode
 3. Fix name input text allowing multiple lines
 4. Fix filling in empty name allowing to proceed to quiz screen
+5. Fix quiz choices not correctly saved
+6. Fix quiz score not saved properly
 
 ## Application Specs